Move all stimuli react components to src/public
