+++ src/components/response/CheckBoxInput.tsx
@@ -1,6 +1,6 @@
 import { Checkbox, Group } from '@mantine/core';
 import { CheckboxResponse } from '../../parser/types';
-import { generateErrorMessage } from '../stimuli/inputcomponents/utils';
+import { generateErrorMessage } from './utils';
 
 type inputProps = {
   response: CheckboxResponse;
