+++ src/components/response/SliderInput.tsx
@@ -1,6 +1,6 @@
 import { Slider, SliderProps, Stack, Text } from '@mantine/core';
 import { SliderResponse } from '../../parser/types';
-import { generateErrorMessage } from '../stimuli/inputcomponents/utils';
+import { generateErrorMessage } from './utils';
 
 type inputProps = {
   response: SliderResponse;
