+++ src/components/response/StringInput.tsx
@@ -1,6 +1,6 @@
 import { TextInput } from '@mantine/core';
 import { ShortTextResponse } from '../../parser/types';
-import { generateErrorMessage } from '../stimuli/inputcomponents/utils';
+import { generateErrorMessage } from './utils';
 
 type inputProps = {
   response: ShortTextResponse;
