+++ src/components/response/TextAreaInput.tsx
@@ -1,6 +1,6 @@
 import { Textarea } from '@mantine/core';
 import { LongTextResponse } from '../../parser/types';
-import { generateErrorMessage } from '../stimuli/inputcomponents/utils';
+import { generateErrorMessage } from './utils';
 
 type inputProps = {
   response: LongTextResponse;
