+++ src/public/brush-interactions/Bar.tsx
