+++ src/public/brush-interactions/Paintbrush.tsx
