+++ src/public/brush-interactions/Scatter.tsx
