+++ src/public/brush-interactions/XAxisBar.tsx
