+++ src/public/brush-interactions/YAxisBar.tsx
