+++ src/public/cleveland/BarChart.tsx
