+++ src/public/cleveland/StackedBarChart.tsx
