+++ src/public/cleveland/chartcomponents/Bars.tsx
