+++ src/public/cleveland/chartcomponents/StackedBars.tsx
