+++ src/public/fairness-jnd/FairnessJND.tsx
@@ -2,8 +2,8 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import { Button, Grid } from '@mantine/core';
 import { IconCircleCheck } from '@tabler/icons-react';
 import { getHotkeyHandler } from '@mantine/hooks';
-import { StimulusParams } from '../../../store/types';
-import { useCurrentStep } from '../../../routes';
+import { StimulusParams } from '../../store/types';
+import { useCurrentStep } from '../../routes';
 
 const Ranking = ({ rankings }: { rankings: number[] }) => {
   return (
